+++ Samples~/Scripts/SequenceFireExample.cs
@@ -1,8 +1,7 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.Events;
-namespace FuzzPhyte.Chain
+
+namespace FuzzPhyte.Chain.Samples
 {
     public class SequenceFireExample : MonoBehaviour
     {